% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_split.R
\name{format_url}
\alias{format_url}
\title{Format URL, splitting to fit in editing window}
\usage{
format_url(length = 80)
}
\arguments{
\item{length}{What is the maximum length of each chunk of the path. Default
is 80}
}
\description{
Format URL, splitting to fit in editing window
}
