% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_split.R
\name{get_info}
\alias{get_info}
\title{Get information on the highlighted path from the editor}
\usage{
get_info(context)
}
\arguments{
\item{context}{Context of the text}
}
\description{
Get information on the highlighted path from the editor
}
\keyword{internal}
