% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_split.R
\name{split_given_length}
\alias{split_given_length}
\title{Split}
\usage{
split_given_length(text, length = 80)
}
\arguments{
\item{text}{Text to be split}

\item{length}{What is the maximum length of each chunk of the path. Default
is 80}
}
\description{
Split
}
